RF: Move MicrophoneDevice to hardware
